server/requirements.txt
appdirs==1.4.4
astroid==2.5.1
black==20.8b1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.7.2
pyserial==3.5
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
